Remove legacy ursm indexing
